DataCrawler/requirements.txt
asgiref==3.2.7
bson==0.5.8
dataclasses==0.6
Django==2.2.12
django-cors-headers==3.2.1
djangorestframework==3.11.0
djongo==1.3.2
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.2.4
azure-functions
requests
